se agrego la recopilación de datos (Es temporal luego se va a borrar)

+++ src/main/java/dev/cerez/tahp/connector/model/OrderResult.java
@@ -1,10 +1,12 @@
 package dev.cerez.tahp.connector.model;
 
+import java.math.BigDecimal;
+
 public record OrderResult(
         String orderId,
-        Double executedQty,
-        Double cumulativeQuoteQty,
-        Double receivedQty
+        BigDecimal executedQty,
+        BigDecimal cumulativeQuoteQty,
+        BigDecimal receivedQty
 ) {
 
 }
